modify

+++ app/controller/society/deleteSociety.js
@@ -9,6 +9,7 @@ function response (res, code, bool, message) {
 }
 
 module.exports = (req, res) => {
+  if (req.user.admin !== true) return response(res, 400, false, 'Accès non autorisé')
   if (req.query.id !== '') {
     db.get().then((db) => {
       db.collection('Society').deleteOne({_id: req.query.id})
